+++ sql/schemas/hapi-core-tables.sql
@@ -21,7 +21,8 @@
 
 CREATE TABLE dataset (
        id INTEGER PRIMARY KEY,
-       code VARCHAR(128) UNIQUE NOT NULL,
+       code VARCHAR(36) UNIQUE NOT NULL,
+       hdx_stub VARCHAR(128) UNIQUE NOT NULL,
        title VARCHAR(1024) NOT NULL,
        provider_code VARCHAR(128) NOT NULL,
        provider_name VARCHAR(512) NOT NULL
@@ -36,7 +37,7 @@ ON dataset(provider_name);
 CREATE TABLE resource (
        id INTEGER PRIMARY KEY,
        dataset_ref INT NOT NULL,
-       code VARCHAR(128) UNIQUE NOT NULL,
+       code VARCHAR(36) UNIQUE NOT NULL,
        filename VARCHAR(256) NOT NULL,
        format VARCHAR(32) NOT NULL,
        update_date DATETIME NOT NULL,