+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - removed *_link fields (they can be constructed in the API layer)
+- dataset and resource codes set to VARCHAR(36) (instead of 128)
+- column hdx_stub added to dataset table
 
 ## [1.1.1] - 2023-09-01